Allow plot_facets to reflect absolute distance
Currently `plot_facets()` uses free scales for each facet, so that all tracks appear to be of a similar size in both x and y dimensions. Here we add a new argument `scales`. This defaults to a value of "free", which gives the current behaviour. With a value of "fixed" the scales are set to "fixed" for each facet, so that longer tracks do actually appear to be larger than tracks that covered less distance.  This requires that all points must be normalised by centering all latitude and longitude values about a mean of zero. This also means that the x and y axes are on the same scale, so the aspect ratio of each track is veridical, rather than each axis being stretched to fill out each facet, which happens with "free" scales.